% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-functional.R
\name{analyze_graph}
\alias{analyze_graph}
\title{Analyze the causal graph to determine constraints and objective}
\usage{
analyze_graph(graph)
}
\description{
The graph must contain edge attributes named "leftside" and "lrconnect"
that takes values 0 and 1. Only one edge may have a value 1 for lrconnect. 
The shiny app returns a graph in this format.
}
