% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{btm_var}
\alias{btm_var}
\title{Recursive function to get the last name in a list}
\usage{
btm_var(x, name = NULL)
}
\arguments{
\item{x}{a list}

\item{name}{name of the top element of the list}
}
\value{
The name of the deepest nested list element
}
\description{
Recursive function to get the last name in a list
}
