\name{causaloptim-package}
\alias{causaloptim-package}
\alias{causaloptim}
\docType{package}
\title{
  An Interface to Specify Causal Graphs and Compute Balke Bounds
}
\description{
  Specify causal graphs using a visual interactive interface and then 
  analyze them and compute symbolic bounds for the causal effects in 
  terms of the observable parameters. 
}
\details{
  Run the shiny app by \code{results <- specify_graph()}. See detailed instructions in the vignett \code{browseVignettes("causaloptim")}. 
}
\author{
Michael C Sachs, Arvid Sjölander, Alexandre Balke, Colorado Reed, and Erin Gabriel

Maintainer: Michael C Sachs <sachsmc@gmail.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  \code{browseVignettes("causaloptim")}
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
