% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-linearity.R
\name{check_linear_constraints}
\alias{check_linear_constraints}
\title{Check linearity of constraint space implied by a set of response functions}
\usage{
check_linear_constraints(respvars, p.vals, prob.form)
}
\arguments{
\item{respvars}{List of response functions as produced by \link{create_response_function}}

\item{p.vals}{Matrix defining which probabilities are observable}

\item{prob.form}{A list with two named elements "out", "cond" where each element is a character vector}
}
\value{
A logical value that is TRUE if the constraint space is linear
}
\description{
Check linearity of constraint space implied by a set of response functions
}
\details{
It is assumed that probabilities of the form p(out | cond) are observed, for each combination of values in p.vals. This is limited currently to DAGs that still fit into out left-side right-side framework
}
\examples{
 ## regular IV case

b <- graph_from_literal(Z -+ X, X -+ Y, Ur -+ X, Ur -+ Y)
V(b)$leftside <- c(1,0,0,0)
V(b)$latent <- c(0, 0,0,1)
V(b)$nvals <- c(2,2,2,2)
E(b)$rlconnect <- E(b)$edge.monotone <- c(0, 0, 0, 0)

graph <- b

observed.variables <- V(graph)[V(graph)$latent == 0]
var.values <- lapply(names(observed.variables), 
                     function(varname) seq(from = 0, to = causaloptim:::numberOfValues(graph, varname) - 1))
names(var.values) <- names(observed.variables)
p.vals <- do.call(expand.grid, var.values)

respvars <- create_response_function(graph, V(b)[2:3], V(b)[1])
prob.form <- list(out = c("Y", "X"), cond = "Z")

check_linear_constraints(respvars, p.vals, prob.form)


}
