% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-parsing.R
\name{const.to.sets}
\alias{const.to.sets}
\title{Translate lists of constraints to lists of vectors}
\usage{
const.to.sets(constr, objterms)
}
\arguments{
\item{constr}{List of constraint terms as character strings}

\item{objterms}{Vector of terms in the objective function}
}
\description{
Translate lists of constraints to lists of vectors
}
