% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constant_term}
\alias{constant_term}
\title{Compute the scalar product of two numeric vectors of the same length}
\usage{
constant_term(numbers1, numbers2)
}
\arguments{
\item{numbers1, numbers2}{Two numeric vectors of the same length.}
}
\value{
A string consisting of the value of the scalar product of \code{numbers1} and \code{numbers2}.
}
\description{
A helper function for \code{\link{evaluate_objective}}.
}
