% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{constraintschangeUI}
\alias{constraintschangeUI}
\title{Dynamically change the UI for input of optional constraints.
This fixes the constraint entered into the constraint text input field, 
once it has been successfully parsed.}
\usage{
constraintschangeUI(constrainttext)
}
\arguments{
\item{constrainttext}{A string representing the user-provided constraint.}
}
\description{
Dynamically change the UI for input of optional constraints.
This fixes the constraint entered into the constraint text input field, 
once it has been successfully parsed.
}
