% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{constraintscheck}
\alias{constraintscheck}
\title{Check that a user-provided constraint is parsable, has valid variables and relations.}
\usage{
constraintscheck(constrainttext, graphres)
}
\arguments{
\item{constrainttext}{A string representing a constraint.}

\item{graphres}{An \code{igraph} object representing a DAG.}
}
\value{
\code{TRUE} if all check pass; else \code{FALSE}.
}
\description{
Check that a user-provided constraint is parsable, has valid variables and relations.
}
