% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{constraintsnamecheck}
\alias{constraintsnamecheck}
\title{Given a DAG, check that the variables in a parsed constraint are in the vertex set of that DAG.}
\usage{
constraintsnamecheck(parsed_constraints, graphres)
}
\arguments{
\item{parsed_constraints}{A data frame representing a parsed constraint 
(as returned by \code{parse_constraints}).}

\item{graphres}{An \code{igraph} object representing a DAG.}
}
\value{
\code{TRUE} if the variable names in\code{parsed_constraints} correspond to ones in \code{V(graphres)}; else \code{FALSE}.
}
\description{
Given a DAG, check that the variables in a parsed constraint are in the vertex set of that DAG.
}
