% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{constraintsoperatorcheck}
\alias{constraintsoperatorcheck}
\title{Check that the relations in parsed constraints are valid, 
i.e., one of "==", "<", ">", "<=", ">=".}
\usage{
constraintsoperatorcheck(parsed_constraints)
}
\arguments{
\item{parsed_constraints}{A data frame representing a parsed constraint 
(as returned by \code{parse_constraints}).}
}
\value{
\code{TRUE} if the realations in\code{parsed_constraints} are valid; else \code{FALSE}.
}
\description{
Check that the relations in parsed constraints are valid, 
i.e., one of "==", "<", ">", "<=", ">=".
}
