% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{constraintsparsecheck}
\alias{constraintsparsecheck}
\title{Check that a user-provided optional constraint is parsable.}
\usage{
constraintsparsecheck(constrainttext, graphres)
}
\arguments{
\item{constrainttext}{A string representing a constraint.}

\item{graphres}{An \code{igraph} object representing a DAG.}
}
\value{
\code{TRUE} if \code{constrainttext} is parsable; else \code{FALSE}.
}
\description{
Check that a user-provided optional constraint is parsable.
}
