% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{constraintUI}
\alias{constraintUI}
\title{Generate HTML for a UI for textual input of optional extra constraints.}
\usage{
constraintUI()
}
\value{
HTML code for a text input field and an action button for parsing.
}
\description{
Generate HTML for a UI for textual input of optional extra constraints.
}
