% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation-modules.R
\name{create_R_matrix}
\alias{create_R_matrix}
\title{Create constraint matrix}
\usage{
create_R_matrix(
  graph,
  obsvars,
  respvars,
  p.vals,
  parameters,
  q.list,
  variables
)
}
\arguments{
\item{graph}{The graph}

\item{obsvars}{Vector of observed variable vertices from the graph}

\item{respvars}{Response function, as returned by \link{create_response_function}}

\item{p.vals}{Observed probability matrix}

\item{parameters}{Vector of ps names}

\item{q.list}{List with q matrices, as returned by \link{create_q_matrix}}

\item{variables}{Vector of qs names}
}
\value{
A list with the R matrix and the string representation
}
\description{
Matrix and text representation of constraints on observed probabilities
}
