% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{cyclecheck}
\alias{cyclecheck}
\title{Check that a given digraph is a DAG, i.e., contains no cycles.}
\usage{
cyclecheck(graphres)
}
\arguments{
\item{graphres}{An \code{igraph} object representing a digraph.}
}
\value{
\code{TRUE} if \code{graphres} is a DAG; else \code{FALSE}.
}
\description{
Check that a given digraph is a DAG, i.e., contains no cycles.
}
