% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{edges_from_input}
\alias{edges_from_input}
\title{Construct a data-frame encoding a digraph from GUI input.}
\usage{
edges_from_input(myin)
}
\arguments{
\item{myin}{Digraph input from GUI.}
}
\value{
A data frame encoding the digraph.
}
\description{
Construct a data-frame encoding a digraph from GUI input.
}
