% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{effectUI}
\alias{effectUI}
\title{Generate HTML for a UI for textual input of causal query.}
\usage{
effectUI(default_effect)
}
\arguments{
\item{default_effect}{A string representing a causal query.}
}
\value{
HTML code for a text field with default value \code{default_effect} 
and an action button for parsing it.
}
\description{
Generate HTML for a UI for textual input of causal query.
}
