% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate_objective}
\alias{evaluate_objective}
\title{Compute the scalar product of a vector of numbers and a vector of both numbers and strings}
\usage{
evaluate_objective(c1_num, p, y)
}
\arguments{
\item{c1_num}{A numeric column matrix.}

\item{p}{A character vector.}

\item{y}{A numeric vector whose length is the sum of the lengths of \code{c1_num} and \code{p}.}
}
\value{
A string consisting of an affine expression in \code{p} corresponding to the scalar product of \code{c(c1_num, p)} with \code{y}.
}
\description{
A helper function for \code{\link{opt_effect}}.
}
