% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-parsing.R
\name{expand_cond}
\alias{expand_cond}
\title{Expand potential outcome conditions}
\usage{
expand_cond(cond, obsnames)
}
\arguments{
\item{cond}{Text string of the condition}

\item{obsnames}{Vector of names of observed variables}
}
\description{
Expand potential outcome conditions
}
