% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-utilities.R
\name{find_cycles}
\alias{find_cycles}
\title{Find cycles in a graph}
\usage{
find_cycles(g)
}
\arguments{
\item{g}{an igraph object}
}
\value{
A list of vectors of integers, indicating the vertex sequences for the cycles found in the graph
}
\description{
Find cycles in a graph
}
