% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{graphres_from_edges}
\alias{graphres_from_edges}
\title{Construct an igraph-object encoding a digraph from a data frame.}
\usage{
graphres_from_edges(edges)
}
\arguments{
\item{edges}{A data frame encoding a digraph, as output by edges_from_input.}
}
\value{
An igraph object encoding the digraph.
}
\description{
Construct an igraph-object encoding a digraph from a data frame.
}
