% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{graphrescheck}
\alias{graphrescheck}
\title{Check that a given digraph satisfies the conditions of 
'no left to right edges', 'no cycles', 'valid number of categories' and 'valid variable names'.
Optionally returns the digraph if all checks are passed.}
\usage{
graphrescheck(graphres, ret = FALSE)
}
\arguments{
\item{graphres}{An \code{igraph} object representing a digraph.}

\item{ret}{A logical value. Default is \code{FALSE}.
Set to \code{TRUE} to also return \code{graphres} if all checks are passed.}
}
\value{
If \code{ret=FALSE} (default): \code{TRUE} if all checks pass; else \code{FALSE}.
If \code{ret=TRUE}: \code{graphres} if all checks pass; else \code{FALSE}.
}
\description{
Check that a given digraph satisfies the conditions of 
'no left to right edges', 'no cycles', 'valid number of categories' and 'valid variable names'.
Optionally returns the digraph if all checks are passed.
}
