% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{helptextUI}
\alias{helptextUI}
\title{Generate HTML for a UI displaying instructional text on method and usage.}
\usage{
helptextUI()
}
\value{
HTML code for a UI with text.
}
\description{
Generate HTML for a UI displaying instructional text on method and usage.
}
