% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-utilities.R
\name{initialize_graph}
\alias{initialize_graph}
\title{Initialize an igraph object for use with causaloptim}
\usage{
initialize_graph(graph)
}
\arguments{
\item{graph}{An object of class igraph}
}
\value{
An igraph with the vertex attributes leftside, latent, and nvals, and edge attributes rlconnect and edge.monotone
}
\description{
Checks for required attributes and adds defaults if missing
}
\examples{
b <- igraph::graph_from_literal(X -+ Y)
b2 <- initialize_graph(b)
V(b2)$nvals

}
