% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-parsing.R
\name{latex_bounds}
\alias{latex_bounds}
\title{Latex bounds equations}
\usage{
latex_bounds(bounds, parameters, prob.sym = "P")
}
\arguments{
\item{bounds}{Vector of bounds as returned by \link{optimize_effect}}

\item{parameters}{The parameters object as returned by \link{analyze_graph}}

\item{prob.sym}{Symbol to use for probability statements in latex, usually "P" or "pr"}
}
\description{
Latex bounds equations
}
