% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linear_expression}
\alias{linear_expression}
\title{Compute the scalar product of a vector of numbers and a vector of strings}
\usage{
linear_expression(numbers, strings)
}
\arguments{
\item{numbers}{A numeric vector.}

\item{strings}{A character vector of the same length as \code{numbers}.}
}
\value{
A string consisting of the corresponding linear combination, including the sign of its first term.
}
\description{
A helper function for \code{\link{evaluate_objective}}.
}
