% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linear_term}
\alias{linear_term}
\title{Compute the product of a single numeric scalar and a single string}
\usage{
linear_term(number, string)
}
\arguments{
\item{number}{A numeric vector of length 1.}

\item{string}{A character vector of length 1.}
}
\value{
A string consisting of the concatenation of \code{number} and \code{string}, including its sign.
}
\description{
A helper function for \code{\link{linear_expression}}.
}
