% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_path}
\alias{list_to_path}
\title{Recursive function to translate an effect list to a path sequence}
\usage{
list_to_path(x, name = NULL)
}
\arguments{
\item{x}{A list of vars as returned by \link{parse_effect}}

\item{name}{The name of the outcome variable}
}
\value{
a list of characters describing the path sequence
}
\description{
Recursive function to translate an effect list to a path sequence
}
\examples{
nofill <- "p{Y(X = 1, M1 = 1, M2(X = 1, M1 = 1)) = 1}"
eff2 <- parse_effect(nofill)$vars[[1]][[1]]
list_to_path(eff2, "Y")

}
