% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-utilities.R
\name{numberOfValues}
\alias{numberOfValues}
\title{Get the number of values of a given variable in the graph}
\usage{
numberOfValues(graph, varname)
}
\arguments{
\item{graph}{An igraph object.}

\item{varname}{A string. The name of a vertex in 'graph'.}
}
\value{
An integer greater than 1. The number of values of 'varname'.
}
\description{
Get the number of values of a given variable in the graph
}
