% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{nvalscheck}
\alias{nvalscheck}
\title{Check that the number of categorical levels for each variable in a graph is at least 2.}
\usage{
nvalscheck(graphres)
}
\arguments{
\item{graphres}{An \code{igraph} object representing a digraph.}
}
\value{
\code{TRUE} if each variable is at least binary; else \code{FALSE}.
}
\description{
Check that the number of categorical levels for each variable in a graph is at least 2.
}
