% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{open_constraintinputUI}
\alias{open_constraintinputUI}
\title{Dynamically open a new UI for textual input of optional additional constraints.}
\usage{
open_constraintinputUI()
}
\description{
Dynamically open a new UI for textual input of optional additional constraints.
}
