% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{open_finalUI}
\alias{open_finalUI}
\title{Dynamically open a new UI with two action buttons; 
one for downloading the results to R and
one for displaying LaTeX code for the bounds.}
\usage{
open_finalUI()
}
\description{
Dynamically open a new UI with two action buttons; 
one for downloading the results to R and
one for displaying LaTeX code for the bounds.
}
