% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{open_latexUI}
\alias{open_latexUI}
\title{Dynamically open a new UI for display of LaTeX code for the computed bounds.}
\usage{
open_latexUI()
}
\description{
Dynamically open a new UI for display of LaTeX code for the computed bounds.
}
