% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{open_newUI}
\alias{open_newUI}
\title{Dynamically open a new UI for causal query entry and parsing, 
as well as UIs for optional extra constraints 
and subsequent display of results.}
\usage{
open_newUI(default_effect)
}
\arguments{
\item{default_effect}{A string representing a causal query.
This parameter is expected to be returned by \code{get_default_effect(graphres)}, 
where \code{graphres} is an \code{igraph} representation of the DAG of interest.
This parameter will be non-empty if in the DAG GUI 
one vertex has been set as the exposure and another as the outcome, 
in which case it will be the total causal risk difference 
comparing the probability that the outcome is \code{1} 
if the exposure had been intervened on to be \code{1} 
against if the exposure had instead been set to \code{0}.}
}
\description{
Dynamically open a new UI for causal query entry and parsing, 
as well as UIs for optional extra constraints 
and subsequent display of results.
}
