% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-optimizer.R
\name{optimize_effect}
\alias{optimize_effect}
\title{Run the Balke optimizer and return bounds as a function}
\usage{
optimize_effect(obj)
}
\description{
Run the Balke optimizer and return bounds as a function
}
