% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-parsing.R
\name{parse_constraints}
\alias{parse_constraints}
\title{Parse text that defines a the constraints}
\usage{
parse_constraints(constraints, obsnames)
}
\arguments{
\item{constraints}{A list of character strings}

\item{obsnames}{Vector of names of the observed variables in the graph}
}
\value{
A data frame
}
\description{
Parse text that defines a the constraints
}
