% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-functional.R
\name{parse_effect}
\alias{parse_effect}
\title{Parse text that defines a causal effect}
\usage{
parse_effect(text)
}
\arguments{
\item{text}{Character string}
}
\value{
A list
}
\description{
Parse text that defines a causal effect
}
