% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{plotServer}
\alias{plotServer}
\title{A Shiny module server to go with \code{plotUI}.
Renders to \code{plotUI} a plot of the DAG of the causal problem.}
\usage{
plotServer(id, graphres)
}
\arguments{
\item{id}{A Shiny module identifier. Should match that of \code{plotUI}.}

\item{graphres}{An \code{igraph} object representing the DAG.}
}
\description{
A Shiny module server to go with \code{plotUI}.
Renders to \code{plotUI} a plot of the DAG of the causal problem.
}
