% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{plotUI}
\alias{plotUI}
\title{Dynamically open a new UI to display a plot of the DAG to be optimized over.}
\usage{
plotUI(id)
}
\arguments{
\item{id}{A Shiny module identifier. Should match that of \code{plotServer}.}
}
\description{
Dynamically open a new UI to display a plot of the DAG to be optimized over.
}
