% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal-model.R
\name{print.causalmodel}
\alias{print.causalmodel}
\title{Print relevant information about the causal model}
\usage{
\method{print}{causalmodel}(x, omit_cf_constraints = FALSE, omit_obs_constraints = FALSE, ...)
}
\arguments{
\item{x}{object of class "causalmodel"}

\item{omit_cf_constraints}{Do not print the counterfactual constraints}

\item{omit_obs_constraints}{Do not print the observable constraints}

\item{...}{Not used}
}
\value{
x, invisibly
}
\description{
Print relevant information about the causal model
}
