% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-utilities.R
\name{print_nvals}
\alias{print_nvals}
\title{Print the number of values of each variable/vertex of the analyzed graph object}
\usage{
print_nvals(graphres)
}
\arguments{
\item{graphres}{an igraph object}
}
\value{
None
}
\description{
Print the number of values of each variable/vertex of the analyzed graph object
}
