% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{querychangeUI}
\alias{querychangeUI}
\title{Dynamically change the UI for entry of causal query.}
\usage{
querychangeUI(effecttext)
}
\arguments{
\item{effecttext}{A string representing a causal query.
This string will be the default value in the text input field for the causal query.}
}
\description{
Dynamically change the UI for entry of causal query.
}
