% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{querycheck}
\alias{querycheck}
\title{Given an admissible causal DAG, check that given a causal query satisfies 
conditions that guarantee the corresponding causal problem to be a linear program.
Throws error messages detailing any conditions violated.}
\usage{
querycheck(effecttext, graphres)
}
\arguments{
\item{effecttext}{A string representing a causal query.}

\item{graphres}{An \code{igraph} object representing a digraph.}
}
\value{
\code{TRUE} if \code{effecttext} is parsable, contains only variables in \code{V(graphres)} 
and satisfies conditions for linearity; else \code{FALSE}.
}
\description{
Given an admissible causal DAG, check that given a causal query satisfies 
conditions that guarantee the corresponding causal problem to be a linear program.
Throws error messages detailing any conditions violated.
}
