% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{queryparsecheck}
\alias{queryparsecheck}
\title{Check that a string representing a causal query can be successfully parsed.}
\usage{
queryparsecheck(effecttext)
}
\arguments{
\item{effecttext}{A string representing a causal query.}
}
\value{
\code{TRUE} if \code{effecttext} can be successfully parsed; else \code{FALSE}.
}
\description{
Check that a string representing a causal query can be successfully parsed.
}
