% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-parsing.R
\name{reduce.sets}
\alias{reduce.sets}
\title{Algebraically reduce sets}
\usage{
reduce.sets(sets)
}
\arguments{
\item{sets}{List of constraints as sets of variables}
}
\description{
Identifies and reduces redundant variables
}
