% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{remove_oldUI}
\alias{remove_oldUI}
\title{Dynamically remove parts of the UI after clicking the 'analyze' action button.}
\usage{
remove_oldUI()
}
\description{
Dynamically remove parts of the UI after clicking the 'analyze' action button.
}
