% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{resultschangeUI}
\alias{resultschangeUI}
\title{Dynamically change the UI for display of results of optimized bounds.}
\usage{
resultschangeUI()
}
\description{
Dynamically change the UI for display of results of optimized bounds.
}
