% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{resultstext}
\alias{resultstext}
\title{Generate a string describing the results of a linear causal problem.}
\usage{
resultstext(resultslist, effecttext)
}
\arguments{
\item{resultslist}{A list of results, as returned by \code{specify_graph()}.}

\item{effecttext}{A string representing a causal query.}
}
\value{
HTML code with text describing the bounds from a linear causal problem.
}
\description{
Generate a string describing the results of a linear causal problem.
}
