% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{resultsUI}
\alias{resultsUI}
\title{Generate HTML for a UI with an action button for computing bounds for a causal problem.}
\usage{
resultsUI()
}
\value{
HTML code with a text message and an action button for computing bounds.
}
\description{
Generate HTML for a UI with an action button for computing bounds for a causal problem.
}
