% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{rlcheck}
\alias{rlcheck}
\title{Check that no edges of a given digraph go from the 'right side' to the 'left side'.}
\usage{
rlcheck(graphres)
}
\arguments{
\item{graphres}{An \code{igraph} object representing a digraph.
This digraph should have at least the binary edge attribute \code{rlconnect}.}
}
\value{
\code{TRUE} if the condition is satisfied; else \code{FALSE}.
}
\description{
Check that no edges of a given digraph go from the 'right side' to the 'left side'.
}
