% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{rlcheck0}
\alias{rlcheck0}
\title{Check that a data frame containing the edges of a digraph with certain attributes 
satisfies the condition of no edges going from the 'right side' to the 'left side'.}
\usage{
rlcheck0(edges)
}
\arguments{
\item{edges}{A data.frame representing a digraph.}
}
\value{
\code{TRUE} if the condition is satisfied; else \code{FALSE}.
}
\description{
Check that a data frame containing the edges of a digraph with certain attributes 
satisfies the condition of no edges going from the 'right side' to the 'left side'.
}
