% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-optimizer.R
\name{shortentxt}
\alias{shortentxt}
\title{Shorten strings to 80 characters wide}
\usage{
shortentxt(x)
}
\arguments{
\item{x}{String}
}
\value{
A string with line breaks to keep the width less than 80 characters
}
\description{
Shorten strings to 80 characters wide
}
