% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-parsing.R
\name{symb.subtract}
\alias{symb.subtract}
\title{Symbolic subtraction}
\usage{
symb.subtract(x1, x2)
}
\arguments{
\item{x1}{First term (subtract from)}

\item{x2}{Second term (subtract)}
}
\description{
Like setdiff but doesn't remove duplicates x1 - x2
}
