% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{vertexnamecheck}
\alias{vertexnamecheck}
\title{Check that the names given to the vertices of a digraph are all valid.}
\usage{
vertexnamecheck(graphres)
}
\arguments{
\item{graphres}{An \code{igraph} object representing a digraph.}
}
\value{
\code{TRUE} if all the variable names are valid; else \code{FALSE}.
}
\description{
Check that the names given to the vertices of a digraph are all valid.
}
