% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causaloptim-package.R
\docType{package}
\name{causaloptim-package}
\alias{causaloptim-package}
\alias{causaloptim}
\title{An Interface to Specify Causal Graphs and Compute Bounds on Causal Effects}
\description{
Specify causal graphs using a visual interactive interface and then analyze them and compute symbolic bounds for the causal effects in terms of the observable parameters.
}
\details{
Run the shiny app by \code{results <- specify_graph()}. See detailed instructions in the vignette \code{browseVignettes("causaloptim")}.
}
\references{
Sachs, M. C., Jonzon, G., Sjölander, A., & Gabriel, E. E. (2023). A general method for deriving tight symbolic bounds on causal effects. Journal of Computational and Graphical Statistics, 32(2), 567-576. \url{https://www.tandfonline.com/doi/full/10.1080/10618600.2022.2071905} .
}
\seealso{
\code{browseVignettes('causaloptim')} \link{specify_graph}
}
\author{
Michael C Sachs, Arvid Sjölander, Gustav Jonzon, Alexander Balke, Colorado Reed, and Erin Gabriel
    Maintainer: Michael C Sachs <sachsmc at gmail.com>
}
