% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal-model.R
\name{create_causalmodel}
\alias{create_causalmodel}
\title{Create a structural causal model from a graph or a set of response functions}
\usage{
create_causalmodel(graph = NULL, respvars = NULL, p.vals, prob.form)
}
\arguments{
\item{graph}{A graph with special edge and vertex attributes, as produced by \link{initialize_graph}}

\item{respvars}{List of response functions as produced by \link{create_response_function}}

\item{p.vals}{Data frame defining which probabilities are observable. The variable names of p.vals must all appear in prob.form}

\item{prob.form}{A list with two named elements "out", "cond" where each element is a character vector of variable names that appear in p.vals}
}
\value{
An object of class "causalmodel"
}
\description{
Create a structural causal model from a graph or a set of response functions
}
\details{
It is assumed that probabilities of the form p(out | cond) are observed, for each combination of values in p.vals. cond may be NULL in which case nothing is conditioned on.
}
\examples{
 ## regular IV case

graph <- initialize_graph(graph_from_literal(Z -+ X, X -+ Y, Ur -+ X, Ur -+ Y))
p.vals <- expand.grid(Z = 0:1, X = 0:1, Y = 0:1)
prob.form <- list(out = c("X", "Y"), cond = "Z")

iv_model <- create_causalmodel(graph, respvars = NULL, p.vals, prob.form)


}
