% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation-modules.R
\name{create_effect_vector}
\alias{create_effect_vector}
\title{Translate target effect to vector of response variables}
\usage{
create_effect_vector(effect, graph, obsvars, respvars, q.list, variables)
}
\arguments{
\item{effect}{Effect list, as returned by \link{parse_effect}}

\item{graph}{The graph}

\item{obsvars}{Vector of observed variable vertices from the graph}

\item{respvars}{Response function, as returned by \link{create_response_function}}

\item{q.list}{List with q matrices, as returned by \link{create_q_matrix}}

\item{variables}{Vector of qs names}
}
\value{
A list with the target effect in terms of qs
}
\description{
Translate target effect to vector of response variables
}
