% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-utilities.R
\name{get_default_effect}
\alias{get_default_effect}
\title{Define default effect for a given graph}
\usage{
get_default_effect(graphres)
}
\arguments{
\item{graphres}{The graph object, should have vertex attributes "outcome" and "exposure"}
}
\value{
A string that can be passed to \link{parse_effect}
}
\description{
Define default effect for a given graph
}
