% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-optimizer.R
\name{interpret_bounds}
\alias{interpret_bounds}
\title{Convert bounds string to a function}
\usage{
interpret_bounds(bounds, parameters)
}
\arguments{
\item{bounds}{The bounds element as returned by \link{optimize_effect}}

\item{parameters}{Character vector defining parameters, as returned by \link{analyze_graph}}
}
\value{
A function that takes arguments for the parameters, i.e., the observed probabilites and returns a vector of length 2: the lower bound and the upper bound.
}
\description{
Convert bounds string to a function
}
