% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_path}
\alias{list_to_path}
\title{Recursive function to translate an effect list to a path sequence}
\usage{
list_to_path(x, name = NULL)
}
\arguments{
\item{x}{A list of vars as returned by parse_effect}

\item{name}{The name of the outcome variable}
}
\value{
a list of characters describing the path sequence
}
\description{
Recursive function to translate an effect list to a path sequence
}
