% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-functional.R
\name{parse_constraints}
\alias{parse_constraints}
\title{Parse text that defines a the constraints}
\usage{
parse_constraints(constraints)
}
\arguments{
\item{constraints}{A list of character strings}
}
\value{
A data frame
}
\description{
Parse text that defines a the constraints
}
