% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-functional.R
\name{plot.linearcausalproblem}
\alias{plot.linearcausalproblem}
\title{Plot the graph from the causal problem}
\usage{
\method{plot}{linearcausalproblem}(x, ...)
}
\arguments{
\item{x}{object of class "linearcausaloptim"}

\item{...}{Not used}
}
\value{
Nothing
}
\description{
Plot the graph from the causal problem
}
