% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-utilities.R
\name{plot_graphres}
\alias{plot_graphres}
\title{Plot the analyzed graph object}
\usage{
plot_graphres(graphres)
}
\arguments{
\item{graphres}{an igraph object}
}
\value{
None
}
\description{
Special plotting method for igraphs of this type
}
