% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-graph.R
\name{print.linearcausalproblem}
\alias{print.linearcausalproblem}
\title{Print the causal problem}
\usage{
\method{print}{linearcausalproblem}(x, ...)
}
\arguments{
\item{x}{object of class "linearcausaloptim"}

\item{...}{Not used}
}
\value{
x, invisibly
}
\description{
Print the causal problem
}
