% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-effect.R
\name{update_effect}
\alias{update_effect}
\title{Update the effect in a linearcausalproblem object}
\usage{
update_effect(obj, effectt)
}
\arguments{
\item{obj}{An object as returned by \link{analyze_graph}}

\item{effectt}{A character string that represents the causal effect of interest}
}
\value{
A object of class linearcausalproblem, see \link{analyze_graph} for details
}
\description{
If you want to use the same graph and response function, but change the 
effect of interest, this can save some computation time.
}
